+++ src/Exception/FastRouteException.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace WoohooLabs\Harmony\Exception;
+
+use Exception;
+
+class FastRouteException extends Exception
+{
+}